% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{sum_vector}
\alias{sum_vector}
\title{Sum of a numeric vector}
\usage{
sum_vector(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric scalar: the sum of x.
}
\description{
Computes the sum of all elements in a numeric vector.
}
\examples{
sum_vector(c(9, 2, 3))
sum_vector(c(NA, 7, 2))  # returns NA if NA present
}
